Updated GeoTIFF documentation.
In particular, added Scaladoc for Encoder, updated GeoTiffWriter
and split some extra classes out into their own files.
