+++ src/main/scala/geotrellis/data/geotiff/Settings.scala
@@ -1,0 +1,8 @@
+package geotrellis.data.geotiff
+
+/**
+ * Used by geotiff.Encoder to control how GeoTiff data is to be written.
+ *
+ * Currently supports sample size and format.
+ */
+case class Settings(size:SampleSize, format:SampleFormat)
